shell is able to be compiled and execute only simple commandes still building tokinizer function to handle commande wtih args or arg+path

+++ string_manage.c
@@ -9,7 +9,7 @@
 * Return: 0 on success
 */
 
-int _strncmp(const char *str1, const char *str2, size_t n)
+int _strncmp(const char *str1, const char *str2, int n)
 {
 	int idx;
